Add files via upload
compute the average of 25 sample IEDs per channel, use it as a template, find the cross correlation with the EEG on other times of IED and find the most correlated aligned IED then plot the final IED average. Then mark the onset and peak of the IED average for each channel and save it.